削除機能の実装要件修正

+++ app/controllers/items_controller.rb
@@ -38,7 +38,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    if user_signed_in? && current_user == item.user
+      item.destroy
+    end
     redirect_to root_path
   end
